requirements.txt
argon2-cffi==19.2.0
arrow==0.15.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
django-model-utils==4.0.0
django-mongoengine==0.3.3
django-mongoengine-filter==0.3.4
django-redis==4.11.0
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangotoolbox==1.8.0
djongo==1.2.38
idna==2.8
itypes==1.2.0
Jinja2==2.10.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
mongo==0.2.0
mongoengine==0.20.0
oauthlib==3.1.0
openapi-codec==1.3.2
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
redis==3.5.0
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.9
